This is the Song Grapher Processing Script
Welcome to the Song Grapher Processing Script.
This Script paints Graphical Songs made from Melodies and them from Events which results in Colors. All this Elements and Layers are connected in a System of Color Intervals.
I hope somebody will like this little Script and plays a little bit with it.