This is the Song Grapher HSB Processing Script
Welcome to the Song Grapher HSB Processing Script.
This is a Further Development of the Song Grapher Program. The Color Model has changed to HSB, and the Calculations of the Delta Values for Color Changes have been changed. Also the Color Mixing has been changed.
I hope somebody will like this little Script-